services/search-service/requirements.txt
fastapi[standard]
uvicorn[standard]
jinja2>=3.1.2
httpx>=0.27.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
alembic>=1.12.0

# Redis Cache
redis>=5.0.0
hiredis>=2.2.0  # Fast Redis protocol parser

# In-Memory Cache
cachetools>=5.3.0  # LRU cache for hot data

# Meilisearch for autocomplete
meilisearch-python-sdk>=3.5.0

# Fuzzy Matching & Search
python-Levenshtein>=0.25.0  # Fast string similarity
rapidfuzz>=3.5.0  # Alternative fuzzy matching

# External APIs
yfinance>=0.2.28
pandas>=2.0.0

# Retry logic & resilience
tenacity>=8.2.0

# Service Discovery
python-consul2>=0.1.5

# Logging & Monitoring
structlog>=23.1.0
python-json-logger>=2.0.7
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-instrumentation-asyncpg>=0.42b0
opentelemetry-exporter-otlp>=1.21.0

# Security
PyJWT>=2.8.0
passlib>=1.7.4

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0
faker>=20.0.0
